Move the tests to the Firmware object
Adding tests to the components seemed like the obvious solution, but I've found
we need to work around the behaviour more than benefit from it.
For instance, scanning a multi-component firmware means decompressing the same
archive multiple times. It's easier to pass with the Firmware object and handle
each Component within the plugin.

While we're breaking the API, rename Assay to Test -- as it means I don't need
to explain what the word means to any more people.

Also, add an optional max_age which allows us to re-process the firmware tests
every few weeks when we have updated malware signatures for example.