Use CHIPSEC to add component shards for UEFI firmware
This allows us to show what modules make up each firmware volume, and in the
future means we can do attestation and verification of the installed modules.

At the moment it's really useful to see what modules are shared between models
and vendors, and would allow us to construct a blocklist in the future for
known broken firmware.