Allow researchers to run YARA queries on the public firmware
Some security researchers want to run YARA queries on the firmware on the LVFS.
They can (and do) this already, but it means:

 * Downloading tens of gigabytes of firmware using PULP_MANIFEST
 * Exploding each cabinet to get the capsule
 * Exploding each capsule into EFI binaries with chipsec
 * Running yara on the command line

Given the LVFS already has the shards cached for each firmware, just allow
authenticated users to run YARA queries on the public firmware. These are done
in an existing cron job as they typically take a few minutes to compile and run.

It's important to note that this functionality isn't providing anything more
than people can do already with wget and a shell script and so I don't think
it's overstepping the mark allowing carefully chosen security researchers
access to this kind of thing. If anything, I can keep an eye on them this way :)