Allow performing YaraQuery scans on component blob
We don't want to be restricted to just scanning shards, and this allows us to
get the same set of matches as the blocklist plugin would.

This means we have to save the Component in the YaraQueryResult and also make
the ComponentShard optional.

This also means we can show components that matched in the Yara query result
list, as in most cases you just care about the models found, not the shards.