Never include firmware in private in any embargo remote
In the instance the firmware went private->embargo->private it should not longer be included even if signed.

This has the side effect that if a version format was changed (e.g. fixed to be the correct value) we could accidentally dedupe the VersionFormat with the old rather than the new value.