Only show vendors with LVFS users on the vendorlist
I don't want people to think I have a relationship with vendors who have just
ignored my emails. It's also probably not a good idea to use the vendor logo
so prominently for known-hostile vendors.

The 'missing' vendors will still be shown if the user searches for it using the
top headerbar.

This also speeds up the vendor list quite considerably.