Do not hide the component certificate page in the nav bar
We have to load every shard, and every certificate of that shard, just to find
out if we should show the nav element. Just show it if there are shards, and
add an infobar in the case the firmware has no parsed certificates.

Fixes the other half of https://github.com/fwupd/lvfs-website/issues/466