Rename the button text in the firmware target view
More than a few people have clicked that button expecting the LVFS to ask where
to move it to, not realising that it would move the firmware _to_ that target.

This is compounded by not showing any buttons when there is no user ACL to move
to testing and stable.