Rewrite the metainfo when signing the firmware
One of the common tripping points for OEMs and ODMs is that the uploaded
firmware.metainfo.xml does not get rewritten if requirements are added or other
changes are made on the LVFS UI.

Although we use the metadata to know what firmware file to download, we use
the 'as uploaded' metadata when actually installing the firmware. By rewriting
the metainfo we can ensure the two information sources match.

This does of course change the downloaded file checksum, but this is not being
served by the CDN.