Allow all users to view the profile page
The profile page no longer just shows the password-change UI, it now allows the
user to upload custom certificates and also opt-in to various notifications for
firmware. At the moment users uing the Azure login method could not get to this
now-useful page.

Remove the ACL for getting to the profile page (even for robots, which may be
operated by a real person to upload things like client certificates...) and use
the correct `@manage-password` ACL when restricting the 2FA feature.