Export the success confidence to the mdsync vendor
The actual number of reports (which will be related to the number of firmware
downloads) is a private stat we can't share with anyone; however it would be
really useful to know the order of magnitude of reports. This way 50% failure
with just two reports would not trigger any alarms.

The high/medium/low states are granular enough to be useful, but should not
leak numbers that OEMs might consider private.