Do not allow users requiring OAuth to use /lvfs/login
If we called /lvfs/login manually using curl, with a username that matched a
Vendor.oauth_domain_glob then we would create a user and the session would be
marked as logged in. Thankfully the schema check on the database for a missing
username saves us, resulting in a uwsgi stacktrace and a 'LVFS is unavailable'
error instead.

The missing code just informs the user that they have to use the *other*
button in the login screen, which I presume was the reason for the call to
`_create_user_for_oauth_username()` in the first place...

Spotted by Justin Steven <justin@justinsteven.com>, many thanks.